# workers/openapi.yaml
openapi: 3.0.0
info:
  title: Email Handler API
  version: 1.0.0
  description: API for managing email handling and routing
servers:
  - url: /api
    description: API base URL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaseDto:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    IncomingHistoryDto:
      allOf:
        - $ref: '#/components/schemas/BaseDto'
        - type: object
          required:
            - fromEmail
            - toEmail
            - subject
          properties:
            fromEmail:
              type: string
              format: email
            toEmail:
              type: string
              format: email
            subject:
              type: string
            destination:
              type: string
            emailClass:
              type: string
              enum: [UNKNOWN, OTP, INVOICE, PROMOTIONAL]
            summary:
              type: string
...
    SettingsDto:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
        value:
          type: string
        description:
          type: string
    SettingsResponseDto:
      allOf:
        - $ref: '#/components/schemas/BaseDto'
        - $ref: '#/components/schemas/SettingsDto'
paths:
...
  /incoming-history:
    get:
      summary: Get incoming email history
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: List of incoming email history
          content:
            application/json:
              schema:
                type: object
                properties:
                  histories:
                    type: array
                    items:
                      $ref: '#/components/schemas/IncomingHistoryDto'

    post:
      summary: Create a new incoming email history record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncomingHistoryDto'
      responses:
        '201':
          description: Incoming email history record created
          content: {}
  /settings:
    get:
      summary: Get all settings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings:
                    type: array
                    items:
                      $ref: '#/components/schemas/SettingsResponseDto'

    put:
      summary: Create or update a setting
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsDto'
      responses:
        '204':
          description: Setting created or updated successfully
          content: {}
  /settings/{key}:
    get:
      summary: Get a setting by key
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Setting details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponseDto'
        '404':
          description: Setting not found

    delete:
      summary: Delete a setting
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Setting deleted
          content: {}